feat(install): worked on using channels for installing servers
So, couldn't get async to directly work the way I wanted, this
is an attempt at getting the install to start off an install
but return right away by using a channel. The install service
now has a run method that watches a receiver and is run
with a liftoff fairing. It compiles and starts up but
I'm not sure if the functionality really works yet